backend/pyproject.toml
[project]
name = "ipes6-backend"
version = "0.1.0"
description = "Backend for IPES6 Pre-registration System"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
requires-python = ">=3.11,<3.13"
dependencies = [
    "Django==5.2.6",
    "django-ninja==1.4.3",
    "django-cors-headers==4.8.0",
    "pymysql", # Recommended for Windows
    "reportlab==4.4.3",
    "qrcode[pil]",
    "Pillow==11.3.0",
    "requests==2.32.5",
    "email-validator==2.3.0",
    "lxml==6.0.1",
    "weasyprint==66.0",
    "pydyf==0.11.0",
    "pyhanko==0.31.0",
    "pyhanko-certvalidator==0.29.0",
    "pypdf==6.0.0",
    "pyphen==0.17.2",
    "python-bidi==0.6.6",
    "arabic-reshaper==3.0.0",
    "fonttools==4.59.2",
    "dnspython==2.8.0",
    "cffi==2.0.0",
    "cryptography==45.0.7",
    "oscrypto==1.3.0",
    "asn1crypto==1.5.1",
    "charset-normalizer==3.4.3",
    "idna==3.10",
    "urllib3==2.5.0",
    "certifi==2025.8.3",
    "Brotli==1.1.0",
    "six==1.17.0",
    "typing-inspection==0.4.1",
    "typing_extensions==4.15.0",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "uritools==5.0.0",
    "webencodings==0.5.1",
    "tinyhtml5==2.0.0",
    "tinycss2==1.4.0",
    "svglib==1.5.1",
    "zopfli==0.2.3.post1",
    "html5lib==1.1",
    "pydantic==2.11.9",
    "pydantic_core==2.33.2",
    "annotated-types==0.7.0",
    "asgiref==3.9.1",
    "sqlparse==0.5.3",
    "python-dotenv>=1.1.1",
]

[project.scripts]
dev = "python manage.py runserver"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false   # 👈 clave: no intentes construir wheel del proyecto local
# (opcional) configs de uv; NO pongas 'python =' acá
# ejemplo:
# python-preference = "only-managed"
